P01-auth/assetarc-auth/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.1
SQLAlchemy==2.0.29
bcrypt==4.1.2
PyJWT==2.8.0
boto3==1.34.158
python-dotenv==1.0.1
email-validator==2.1.1
